refactor to move form to separate class
